refactor(mcp-server): migrate types and utils to @disclaude/mcp-server (Issue #1042)
Migrates MCP tool types and utility functions from src/mcp/ to the
@disclaude/mcp-server package as part of the architecture refactoring.

Changes:
- Add MCP tool types (SendMessageResult, SendFileResult, etc.) to mcp-server
- Add Feishu API utilities (sendMessageToFeishu, replyInThread, etc.)
- Add card validation utilities (isValidFeishuCard, getCardValidationError)
- Update src/mcp files to re-export from @disclaude/mcp-server for backward compatibility
- Add @larksuiteoapi/node-sdk dependency to mcp-server package

This is the first step in separating MCP Server code to its own package.
Subsequent PRs will migrate the remaining MCP tools and server code.

Related: #1037 (Architecture refactoring parent issue)
