+++ packages/mcp-server/src/tools/index.ts
@@ -1,0 +1,16 @@
+/**
+ * MCP tool types.
+ *
+ * @module mcp-server/tools
+ */
+
+export type {
+  SendMessageResult,
+  SendFileResult,
+  MessageSentCallback,
+  ActionPromptMap,
+  InteractiveMessageContext,
+  SendInteractiveResult,
+  AskUserOptions,
+  AskUserResult,
+} from './types.js';
