+++ packages/mcp-server/src/utils/index.ts
@@ -1,0 +1,23 @@
+/**
+ * MCP utilities.
+ *
+ * @module mcp-server/utils
+ */
+
+export { isValidFeishuCard, getCardValidationError } from './card-validator.js';
+
+export {
+  sendMessageToFeishu,
+  replyInThread,
+  getThreads,
+  getThreadMessages,
+} from './feishu-api.js';
+
+export type {
+  SendMessageResult as FeishuSendMessageResult,
+  ReplyInThreadResult,
+  ThreadItem,
+  GetThreadsResult,
+  ThreadMessageItem,
+  GetThreadMessagesResult,
+} from './feishu-api.js';
