+++ src/mcp/utils/index.ts
@@ -2,7 +2,15 @@
  * Utility functions for Feishu MCP.
  *
  * @module mcp/utils
+ *
+ * @deprecated Import from @disclaude/mcp-server instead.
+ * This file re-exports from @disclaude/mcp-server for backward compatibility.
  */
 
-export { isValidFeishuCard, getCardValidationError } from './card-validator.js';
-export { sendMessageToFeishu } from './feishu-api.js';
+export { isValidFeishuCard, getCardValidationError } from '@disclaude/mcp-server';
+export {
+  sendMessageToFeishu,
+  replyInThread,
+  getThreads,
+  getThreadMessages,
+} from '@disclaude/mcp-server';
